Move helper modules into a utilities directory
