+++ lib/active_record_extended.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require "active_record_extended/version"
-require "active_record_extended/utilities"
-require "active_record_extended/order_by_utilities"
+require "active_record_extended/utilities/support"
+require "active_record_extended/utilities/order_by"
 require "active_record_extended/active_record"
 require "active_record_extended/arel"
